requirements.txt
black
discord.py
httpx
mkdocs-linkcheck
mkdocs-material
mkdocstrings[python]
mypy
openai>=1.0.0
pre-commit
pytest
pytest-asyncio
pytest-cov
python-dotenv
rich
ruff
tenacity
tiktoken
